% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_results}
\alias{example_results}
\title{Example results using the example data set}
\format{A \code{data.frame} with 1000 rows and 9 variables (columns):
\describe{
\item{Loc}{ordinal number of SNP-CpG pair being analyzed}
\item{numIDV}{number of observations of SNP-CpG pair being analyzed}
\item{beta}{the fixed effect parameter estimate for the predictor of interest}
\item{se_beta}{the standard deviation of fixed effect}
\item{pvalue}{P-value for the fixed effect, based on the wald test}
\item{h2}{heritability of the transformed rate}
\item{sigma}{total variance component}
\item{converged}{a logical indicator for convergence}
\item{time}{time to converge}
}}
\usage{
example_results
}
\description{
A data set containing the output of \code{image} function call on the
included data set \code{ExampleData}.
}
\keyword{datasets}
