\name{IMAGE-package}
\alias{IMAGE-package}
\alias{IMAGE}
\docType{package}
\title{\packageTitle{IMAGE}}
\description{\packageDescription{IMAGE}mQTL mapping in bisulfite sequencing studies by fitting a binomial mixed model, incorporating allelic-specific methylation pattern.}

\author{
  Yue Fan, Shiquan Sun,  Xiang Zhou
  Maintainer: Yue Fan <yuef@umich.edu>
}

\keyword{package}


