% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMAGE.R
\name{image}
\alias{image}
\title{High-powered detection of genetic effects on DNA methylation using integrated methylation QTL mapping and allele-specific analysis.}
\usage{
image(geno, data, K, Covariates = NULL, numCore = 1,
  fit.maxiter = 500, fit.tol = 1e-05, verbose = TRUE)
}
\arguments{
\item{geno}{a data list containing the genotype data.}

\item{data}{a data list containing the methylation data.}

\item{K}{a known kinship matrix. This matrix should be a positive semi-definite matrix with dimensions equal to the samplie size.}

\item{Covariates}{a matrix containing the covariates subject to adjustment (Default = NULL).}

\item{numCore}{a positive integer specifying the number of cores for parallel computing (default = 1).}

\item{fit.maxiter}{a positive integer specifying the maximum number of iterations when fitting the generalized linear mixed model (default = 500).}

\item{fit.tol}{a positive number specifying tolerance, the difference threshold for parameter estimates below which iterations should be stopped (default = 1e-5).}

\item{verbose}{a logical switch for printing detailed information (parameter estimates in each iteration) for testing and debugging purpose (default = TRUE).}
}
\value{
\item{loc}{ordinal number of SNP-CpG pair being analyzed}

\item{numIDV}{number of observations of SNP-CpG pair being analyzed}

\item{beta}{the fixed effect parameter estimate for the predictor of interest.}

\item{se_beta}{the standard deviation of fixed effect.}

\item{pvalue}{P value for the fixed effect, based on the wald test.}

\item{h2}{heritability of the transformed rate.}

\item{sigma2}{total variance component.}

\item{converged}{a logical indicator for convergence.}
}
\description{
mQTL mapping in bisulfite sequencing studies by fitting a binomial mixed model, incorporating allelic-specific methylation pattern.
}
\details{
IMAGE properly accounts for the count nature of bisulfite sequencing data and incorporates allele-specific methylation patterns from heterozygous individuals to enable more powerful mQTL discovery. 
Binomial mixed models (BMM) are fitted using the penalized quasi-likelihood (PQL) method proposed by Breslow and Clayton (1993).
}
\examples{
data(exampledata)
attach(exampledata)
res=image(geno,data,K)
closeAllConnections()
detach(exampledata)
}
\references{
Breslow, N.E. and Clayton, D.G. (1993) Approximate Inference in Generalized Linear Mixed Models. Journal of the American Statistical Association 88, 9-25.
}
\author{
Yue Fan, Shiquan Sun, Xiang Zhou
}
