\name{shift}
\alias{shift}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Shift A Vector}
\description{  Backward shift i units of a vector . }
\usage{shift(x,i=1)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{The vector to be shifted}
  \item{i}{The number of units wanted to be shifted. Default value is 1.}
}
%\details{
%  
%}
%\value{
%  
%}
%\references{
%% ~put references to the literature/web site here ~
%}
%\author{
%%  ~~who you are~~
%}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
shift(1:5,2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{utilities}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
