% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example data set to test the IMAGE method}
\format{A \code{list} three elements:
\describe{
\item{data}{A \code{data.frame} with columns r, y, r1, r2, y1, y2}
\item{geno}{A \code{data.frame} with columns hap1 and hap2}
\item{K}{A 100X100 matrix}
}}
\usage{
ExampleData
}
\description{
A list containing the data needed to use the \code{image} function.
}
\keyword{datasets}
