% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example dataset}
\format{Contains the following objects:
\describe{
  \item{geno}{a data list containing 2 haplotypes.}
  \item{data}{a data list containing the methylated read counts and total read counts for each allele.}
  \item{K}{a genetic relationship matrix.}
}}
\usage{
exampledata
}
\description{
A simulated example dataset of IMAGE.
The variables are as follows:
}
\examples{
data(exampledata)
res=image(exampledata$geno,exampledata$data,exampledata$K)
}
\keyword{datasets}
