\name{flash-package}
\alias{flash-package}
\alias{flash}
\docType{package}
\title{
Preprocess pictures and then do inference to detect flash area
}
\description{
Haven't done yet. Wait for updated
}
\details{
\tabular{ll}{
Package: \tab flash\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## This part must be written in valid R syntax!
}
